Move functions in their own unit file.
I' have moved several functions in new files with names consisting of the original fike name suffixed by the fucntion name.
The goal is to allow the linker to eliminate dead code.
Plesae note that this work is very partial: I made it only on files of interest for my current project, but this can already help people to reduce code size.